Chart Display: Divide JFreeChart depended and independed stuff into different packages.
