+++ src/main/java/net/imagej/plot/MarkerStyle.java
@@ -1,8 +1,5 @@
 package net.imagej.plot;
 
-/**
- * Created by arzt on 14/12/2016.
- */
 public enum MarkerStyle {
 	NONE, PLUS, X, STAR, SQUARE, FILLEDSQUARE, CIRCLE, FILLEDCIRCLE
 }
