+++ src/main/java/net/imagej/plot/PlotService.java
@@ -1,10 +1,11 @@
 package net.imagej.plot;
 
 import net.imagej.ImageJService;
-import net.imagej.table.Table;
+import net.imagej.ui.swing.viewer.plot.jfreechart.JfcScatterPlot;
 
-// TODO: consider extending WrapperService, and making Plot into a WrapperPlugin
+// TODO: consider extending WrapperService, and making JfcPlot into a WrapperPlugin
 public interface PlotService extends ImageJService {
-	BoxPlot createBoxPlot();
+
 	ScatterPlot createScatterPlot();
+
 }
