+++ src/main/java/net/imagej/plot/SeriesStyle.java
@@ -1,45 +1,18 @@
 package net.imagej.plot;
 
-import org.jfree.data.xy.XYSeries;
-
 import java.awt.Color;
 
-/**
- * Created by arzt on 14/12/2016.
- */
+public interface SeriesStyle {
 
-public class SeriesStyle {
-	private Color color;
-	private MarkerStyle marker_style;
-	private LineStyle line_style;
+	void setColor(Color color);
 
-	SeriesStyle() {
-		color = null;
-		line_style = null;
-		marker_style = null;
-	}
+	void setLineStyle(LineStyle style);
 
-	public void setColor(Color color) {
-		this.color = color;
-	}
+	void setMarkerStyle(MarkerStyle style);
 
-	public void setLineStyle(LineStyle style) {
-		this.line_style = style;
-	}
+	Color getColor();
 
-	public void setMarkerStyle(MarkerStyle style) {
-		this.marker_style = style;
-	}
+	LineStyle getLineStyle();
 
-	public Color getColor() {
-		return color;
-	}
-
-	public LineStyle getLineStyle() {
-		return line_style;
-	}
-
-	public MarkerStyle getMarkerStyle() {
-		return marker_style;
-	}
+	MarkerStyle getMarkerStyle();
 }
