+++ src/main/java/net/imagej/ui/swing/viewer/plot/jfreechart/PlotDisplayViewer.java
@@ -8,13 +8,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -29,23 +29,20 @@
  * #L%
  */
 
-package net.imagej.plot;
+package net.imagej.ui.swing.viewer.plot.jfreechart;
 
-import org.scijava.display.AbstractDisplay;
-import org.scijava.display.Display;
-import org.scijava.plugin.Plugin;
+import net.imagej.ui.swing.viewer.plot.jfreechart.JfcPlot;
+import net.imagej.ui.swing.viewer.plot.jfreechart.PlotDisplay;
+import org.scijava.ui.viewer.DisplayViewer;
 
 /**
- * Default display for {@link Plot}s.
- *
+ * A display viewer for a display of {@link JfcPlot}s.
+ * 
  * @author Curtis Rueden
  */
-@Plugin(type = Display.class)
-public class DefaultPlotDisplay extends AbstractDisplay<Plot> implements
-	PlotDisplay
-{
+public interface PlotDisplayViewer extends DisplayViewer<JfcPlot> {
 
-	public DefaultPlotDisplay() {
-		super(Plot.class);
-	}
+	@Override
+	PlotDisplay getDisplay();
+
 }
